+++ CHANGES.md
@@ -1,7 +1,8 @@
 # Version 2.0.1 (to be released)
 
 * Added ToStringTransformer in core
-* Added PropertyChangeTrigger in core
+* Added PropertyChangeTrigger in core that can be used as PropertyChangeListener to trigger the validation.
+* Added ActionTrigger in Swing support that can be used as an Action or ActionListener to trigger the validation.
 * Fixed compilation warning when adding rule input transformer using the GeneralValidatorBuilder
 * Added comments